pipeline propagation preperations
