+++ frontend/src/hooks/explore/useEventHandlers.ts
@@ -1,17 +1,16 @@
-import { useRef } from 'react';
-import { NodeId } from '~/types/explore/nodeTypesCategories';
-// Relative imports here to avoid the ESLint warning.
-// These should never be imported by themselves in another file.
 import { useConnections } from './useConnections';
 import { useDragDrop } from './useDragDrop';
 import { useNodeOperations } from './useNodeOperations';
 
+/**
+ * This handler handles any user-related pipeline interactions.
+ * The sub-handlers such as useNodeOperations() should never be imported on their own.
+ * Always import handlers over the useEventHandlers.
+ */
 export const useEventHandlers = () => {
-    const directedNeighborMap = useRef(new Map<NodeId, NodeId[]>());
+    const { onNodeDataChange, onNodeDelete, onNodesChange } = useNodeOperations();
 
-    const { onNodeDataChange, onNodeDelete, onNodesChange } = useNodeOperations(directedNeighborMap);
-
-    const { handleConnect, isValidConnection, onEdgeDelete } = useConnections(directedNeighborMap);
+    const { handleConnect, isValidConnection, onEdgeDelete } = useConnections();
 
     const { onDragOver, onDrop } = useDragDrop(onNodeDataChange);
