+++ frontend/src/stores/slices/createHistogramSlice.ts
@@ -1,0 +1,14 @@
+import { StateCreator } from 'zustand';
+import { ExploreFlowStore, HistogramSlice } from '~/types/store/exploreStore.types';
+
+export const createHistogramSlice: StateCreator<ExploreFlowStore, [], [], HistogramSlice> = (set) => ({
+    histogramStates: {},
+    setHistogramState: (nodeId, state) => {
+        set((prev) => ({
+            histogramStates: {
+                ...prev.histogramStates,
+                [nodeId]: state,
+            },
+        }));
+    },
+});
